build tag 0.03

+++ build.sh
@@ -2,17 +2,19 @@
 export PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin
 FROM="alpine:3.13"
 VERSION=0.03
-IMAGE=nc_bench2
+IMAGE=nextcloud_benchmark_v2
 TO="christian773/$IMAGE"
 
 
 cd $(dirname $0) 
 git tag -a v$VERSION -m "build tag $VERSION"
-docker system prune -a -f
 git commit -a -m "build tag $VERSION"
 git push
 git push --tags
 
+
+exit 0 # run this if you want to build your own images
+docker system prune -a -f
 sed -i "s@^FROM .*@FROM $FROM@" Dockerfile
 sed -i "s@^ARG VERSION=.*@ARG VERSION=$VERSION@" Dockerfile
